requirements.txt
pdfplumber==0.10.3
PyPDF2==3.0.1
openai>=1.6.1
langchain==0.0.350
langchain-openai==0.0.2
chromadb==0.4.18
sentence-transformers>=2.2.2
python-dotenv==1.0.0
streamlit==1.28.1
pandas==2.1.4
numpy>=1.26.0
huggingface-hub>=0.19.0
ctransformers>=0.2.27
llama-cpp-python>=0.2.11
google-generativeai
# OCR dependencies for scanned PDFs
pytesseract
Pillow
pdf2image 